move account name logic into custom provider
