+++ home/utils/AccountNameContext.tsx
@@ -1,11 +1,12 @@
-import { createContext, useContext } from 'react';
+import React, { useState, createContext, useContext, useEffect } from 'react';
+import AsyncStorage from '@react-native-async-storage/async-storage';
 
 export type AccountNameContextValue = {
   accountName?: string;
   setAccountName: (accountName?: string) => void;
 };
 
-export const AccountNameContext = createContext<AccountNameContextValue | null>(null);
+const AccountNameContext = createContext<AccountNameContextValue | null>(null);
 
 export function useAccountName() {
   const context = useContext(AccountNameContext);
@@ -15,4 +16,24 @@ export function useAccountName() {
   }
 
   return context;
+}
+
+export function AccountNameProvider({ children }: { children: React.ReactNode }) {
+  const [accountName, setAccountName] = useState<string | undefined>();
+
+  useEffect(
+    // when a user changes what account they are viewing as, we should save that preference so when they come back to the app, they see the same account
+    function persistCurrentAccount() {
+      if (accountName) {
+        AsyncStorage.setItem('currentAccount', accountName);
+      }
+    },
+    [accountName]
+  );
+
+  return (
+    <AccountNameContext.Provider value={{ accountName, setAccountName }}>
+      {children}
+    </AccountNameContext.Provider>
+  );
 }
